server/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
google-api-core==2.17.1
google-api-python-client==2.124.0
google-auth==2.28.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonref==1.1.0
kiwisolver==1.4.8
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.66.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.9
rsa==4.9
scipy==1.15.2
sentencepiece==0.2.0
shellingham==1.5.4
sieve==0.1.9
sievedata==1.4.10
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sseclient==0.0.27
textblob==0.19.0
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
youtube-transcript-api==1.0.2
yt-dlp==2025.2.19
